Issue #67: move a counter logic test into a separate module
