+++ fencer.cabal
@@ -110,6 +110,7 @@ test-suite test-fencer
   hs-source-dirs:
     test
   other-modules:
+    Fencer.Logic.Test
     Fencer.Types.Test
     Fencer.Rules.Test
     Fencer.Server.Test
